// packages/learning-service/src/courses/career-level.enum.ts
export enum CareerLevel {
  PROFESSIONAL_II = 'Professional II',
  PROFESSIONAL_III = 'Professional III',
  PROFESSIONAL_IV = 'Professional IV',
  MANAGER_I = 'Manager I',
  MANAGER_II = 'Manager II',
  DIRECTOR_I = 'Director I',
  DIRECTOR_II = 'Director II',
  DIRECTOR_III = 'Director III',
  DIRECTOR_IV = 'Director IV',
}